requirements.txt
Flask==2.3.2
xgboost==2.0.2
azure-storage-blob==12.13.1
numpy==1.23.3
pandas==1.4.3
